The script now ensures folders exist before purging

+++ minecraft/scripts/purge-old.bash
@@ -26,20 +26,20 @@ cd "$CWD"
 ## Do the thing.
 declare -i I=0
 while [[ I -lt ${#DIRS[@]} && I -lt ${#EXTS[@]} ]]; do
+    if [[ -d "${DIRS[I]}" ]]; then
 
-    ## Go to the target directory and print a summary of what files will be deleted.
-    cd "${DIRS[I]}"
-    printf "\n\033[1m$(pwd)/*${EXTS[I]}\033[0m\n"
+        ## Go to the target directory and print a summary of what files will be deleted.
+        cd "${DIRS[I]}"
+        printf "\n\033[1m$(pwd)/*${EXTS[I]}\033[0m\n"
 
-    ## Delete files older than 7 days.
-    find . -type f -mtime +7 |\
-    grep "\\${EXTS[I]}$" |\
-    xargs rm -fv
+        ## Delete files older than 7 days.
+        find . -type f -mtime +7 |\
+        grep "\\${EXTS[I]}$" |\
+        xargs rm -fv
 
-    ## Done.
-    cd "$CWD"
+        ## Done.
+        cd "$CWD"
+    fi
     I=$I+1
-
-## Done-done.
 done
 exit 0
